pyproject.toml
[tool.poetry]
name = "final-pyweb"
version = "0.1.0"
description = ""
authors = ["Marko Kojic <kojicmarko96@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
sqlalchemy = "^2.0.27"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
psycopg = {extras = ["binary"], version = "^3.1.18"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["argon2"], version = "^1.7.4"}
types-python-jose = "^3.3.4.20240106"
types-passlib = "^1.7.7.20240106"
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
ruff = "^0.2.1"
pytest = "^8.0.0"
httpx = "^0.26.0"
pre-commit = "^3.6.1"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
lint.select = [
  "F",     # pyflakes
  "E",     # pycodestyle errors
  "W",     # pycodestyle warnings
  "I001",  # isort
  "C4",    # flake8-comprehensions
  "B",     # flake8-bugbear
  "ISC",   # flake8-implicit-str-concat
  "T20",   # flake8-print
]
lint.ignore = ["ISC001"]

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.params.Depends", "fastapi.Query", "fastapi.params.Query"]

[tool.mypy]
strict = true
plugins = ["sqlalchemy.ext.mypy.plugin"]

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning:passlib.handlers.argon2",
]